replaced restart function, to something a bit more documented and more "palatable" refactored some corestuffs prep for configs and savings of those, also dependency injection... (should probably have led with that ;) )
