+++ SystemT00ls.CoreFunctions/IntTools.cs
