+++ SystemT00ls.CoreFunctions/appsettings.json
@@ -1,0 +1,2 @@
+{
+}
