+++ SystemT00ls/SystemT00ls.csproj
@@ -35,6 +35,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.8" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\EasyConsole.Core\EasyConsole.Core\EasyConsole.Core.csproj" />
     <ProjectReference Include="..\SystemT00ls.CoreFunctions\SystemT00ls.CoreFunctions.csproj" />
   </ItemGroup>
